+++ Implem.Pleasanter/Models/Wikis/WikiModel.cs
@@ -1156,7 +1156,7 @@ namespace Implem.Pleasanter.Models
             {
                 if (notification.HasRelatedUsers())
                 {
-                    var users = new List<long>();
+                    var users = new List<int>();
                     Repository.ExecuteTable(
                         context: context,
                         statements: Rds.SelectWikis(
@@ -1169,8 +1169,8 @@ namespace Implem.Pleasanter.Models
                                 .AsEnumerable()
                                 .ForEach(dataRow =>
                                 {
-                                    users.Add(dataRow.Long("Creator"));
-                                    users.Add(dataRow.Long("Updator"));
+                                    users.Add(dataRow.Int("Creator"));
+                                    users.Add(dataRow.Int("Updator"));
                                 });
                     notification.ReplaceRelatedUsers(
                         context: context,
